Fix unintentional double hooking of GetDeviceState in some case
GetDeviceState was previously hooked once per unique device GUID. This is problematic since different devices can share the same vtable. In cases where vtables are shared we hooked twice resulting in an infinite loop when invokeOriginal is called on the duplicate hook. This patch changes the hooking strategy to hook once per vtable.